#2 Adds unit tests for problematic input.
The issue here is the omitted trailing semicolon, followed by no
whitespace (other than the newline) and then the closing
brace. CSSMinifier assumes it can safely truncate by two characters on
the trailing end, because the first will be the brace, followed by the
semicolon. And if there's whitespace instead, the issue doesn't
manifest. In these cases, though, it truncates a character from the
property's content.